requirements.txt
streamlit==0.66.0
tensorflow-cpu==1.15
numpy==1.17.2
pandas==0.24.2
pillow==7.2.0
keras==2.3.0
opencv-contrib-python-headless
matplotlib
